+++ phpBB/install/database_update.php
@@ -1354,12 +1354,15 @@ function add_bots()
 			WHERE bot_id $sql_id";
 		$db->sql_query($sql);
 
-		$_tables = array(USERS_TABLE, USER_GROUP_TABLE);
-		foreach ($_tables as $table)
+		if (sizeof($user_id_ary))
 		{
-			$sql = "DELETE FROM $table
-				WHERE " . $db->sql_in_set('user_id', $user_id_ary);
-			$db->sql_query($sql);
+			$_tables = array(USERS_TABLE, USER_GROUP_TABLE);
+			foreach ($_tables as $table)
+			{
+				$sql = "DELETE FROM $table
+					WHERE " . $db->sql_in_set('user_id', $user_id_ary);
+				$db->sql_query($sql);
+			}
 		}
 	}
 	else
